// src/shared/types/user.types.ts
export enum RoleEnum {
  ADMIN = 'admin',
  USER = 'user',
  SELLER = 'seller',
}

export enum GenderEnum {
  MALE = 'male',
  FEMALE = 'female',
}

export enum ProviderEnum {
  GOOGLE = 'google',
  SYSTEM = 'system',
}

export enum OtpTypeEnum {
  CONFIRMEMAIL = 'confirm email',
  RESETPASSWORD = 'reset password',
}